remove AssociativeBoth instances where IdentityBoth is available (#289)
* remove AssociativeBoth instances where IdentityBoth is available

* move instances from IdentityBoth to AssociativeBoth

* remove unused imports from IdentityBoth

* move instances from IdentityEither to AssociativeEither

* trigger build
