+++ src/main/scala/zio/prelude/Ord.scala
@@ -194,15 +194,6 @@ object Ord extends Lawful[Ord] {
     (l: A, r: A) => Ordering.fromCompare(ordering.compare(l, r))
 
   /**
-   * The `AssociativeBoth` instance for `Ord`.
-   */
-  implicit val OrdAssociativeBoth: AssociativeBoth[Ord] =
-    new AssociativeBoth[Ord] {
-      def both[A, B](fa: => Ord[A], fb: => Ord[B]): Ord[(A, B)] =
-        fa.both(fb)
-    }
-
-  /**
    * The `Contravariant` instance for `Ord`.
    */
   implicit val OrdContravariant: Contravariant[Ord] =
